Fixed RecipeFormatter::padToLength
Previously attempted to negate an unsigned int (I'm not sure what that does) when generating padded string. Now static_casts the length int to an int, allowing proper negation (and therefore left-aligned text).

+++ src/RecipeFormatter.cpp
@@ -591,7 +591,7 @@ unsigned int RecipeFormatter::getMaxLength( QStringList* list )
 
 QString RecipeFormatter::padToLength( const QString &str, unsigned int length )
 {
-   return QString("%1").arg(str, -length, ' ');
+   return QString("%1").arg(str, -static_cast<int>(length), ' ');
 }
 
 void RecipeFormatter::padAllToMaxLength( QStringList* list, unsigned int padding )
